RF/OPT: just assign None to linux_distribution*, and False to on_debian_wheezy
Those are too expensive to assign a proper value on recent Pythons where
we have to use distro (see https://github.com/datalad/datalad/issues/4690).

on_debian_wheezy was not used anywere in this code base and cannot be True these days
anyways since wheezy EOLed 2 years back, and we do not even provide backports of datalad
for it from neurodebian.

The other variables were used only for a better git-annex installation
suggestion message -- we might want to get away from that and just refer to
installation instructions anyways, but or now I just kept it around.

Even though these changes can possibly considered a "breakage", I think the
actual breakage of anything is very unlikely since code should not really rely in any
critical functionality on what distribution and its release it is running on.

I had kept variables with None values to really avoid any breakage, and so we could in
0.14 announce their proper deprecation cycle and remove them later on

Closes #4690