BF/TST: Respect tmp dir settings
We didn't properly respect DATALAD_TESTS_TEMP_DIR where we should and
it should default to TMPDIR. Furthermore, we only respecteed it as an
env var, not as a config item. And finally, we did use its value where
we shouldn't. That is in datalad/utils. But it's explicitly saying it's
about datalad's tests. Actual datalad code outside the tests should
rather respect TMPDIR instead.