BF: create-sibling: Avoid overwriting existing web log file
The post-update hook installed by `create-sibling --ui ...` constructs
a log file name using a second-resolution `date` call.  For normal
use, consecutive updates are unlikely to result in a file name
collision, but it can very easily happen in a test setting.

These collisions are very likely behind the intermittent failures that
have been happening in test_replace_and_relative_sshpath, where the
number of log files is coming up one short of the expected number
after a call to publish().

Check if the target file exists and, if needed, append a number.

Fixes #4741.