TST: travis: Test with Git's master in _DL_UPSTREAM_GIT cron job
tools/ci/install-latest-git.sh installs the latest Git available in
the ppa:git-core/ppa repository (which tracks the latest release
closely), exiting early if that Git version happens to match the
bundled version.  However, we recently switched to testing with a
git-annex from conda by default, so the latest cron job failed in the
setup phase because a bundled git wasn't present [0].

One option would be to update the script to not expect a bundled git
to exist.  However, instead let's just update the cron job to use the
master branch from git.git.  This allows us to catch issues sooner
than waiting for a release.  The downside is that we may run into
issues that won't actually make it into a Git release.  But (1) Git's
master branch is usually quite stable indicator of what makes will it
into a release because topic branches first have to graduate from next
and (2) I volunteer for the task of distinguishing "in release or in
master-only" failures.

Since we're doing a full rewrite of the script anyway, rename it to be
more consistent with the Travis option.  Also, note that, although
this script is short and simple, it is similar enough to
tools/ci/install-minimum-git.sh in the master branch that it might be
worth merging the two.

[0]: https://travis-ci.org/github/datalad/datalad/jobs/709288940