TST: test_annex_ssh: Avoid syncing unrelated histories
With _DL_ANNEX_INSTALL_SCENARIO="conda-forge 7.20190819" on Travis and
a Docker SSH target, the very last assertion in test_annex_ssh() is
failing [0]: ok_(exists(socket_2))

If Travis is switched to using neurodebian (7.20190819) or the
autobuild, test_annex_ssh does not fail [1, 2].  It also doesn't fail
if I use install-annex.sh to install "conda-forge 7.20190819" locally
with the Docker SSH target.  In a Xenial VM with the "conda-forge
7.20190819" installed (called as it is in .travis.yml), I'm able to
trigger that assertion to fail, kind of (more details at the end).

The sync calls in these tests are a bit odd because they are for
remotes with unrelated histories, so the calls are expected to fail,
but for the purposes testing that a socket is created, that shouldn't
matter.  Switch the test over to a more typical setup, with a local
repository and two remotes that have related histories.  For the
socket-creating annex operation, do a git-annex-copy call that is
expected to succeed.  This makes test_annex_ssh pass on Travis [3] and
in the Xenial VM.  It also speeds the test up (1.5s to 11.5s,
comparing one run each on my local machine), presumably due mostly to
switching away from with_testrepos().

[0] https://travis-ci.org/github/datalad/datalad/jobs/710116798#L1983
[1] https://travis-ci.org/github/datalad/datalad/builds/710151357
[2] https://travis-ci.org/github/datalad/datalad/builds/710165830
[3] https://travis-ci.org/github/datalad/datalad/builds/711536353

----------------------------------------------------------------------

More details on Xenial VM failure.  test_annex_ssh() actually failed
with the "socket 1 exists" check after the _first_ sync call, not the
socket 2 check after the second sync call.  To try to get a better
understanding of what is failing, I removed the first sync call from
the swallow_outputs() context.  But then the socket 1 test didn't fail
while the socket 2 check at the end did fail.  Trying several times, I
didn't notice any flakiness to this.

Anyway, I have no real understanding of the observations above, or why
the sync call is failing in general, or why it happens only on Xenial
with the "conda-forge 7.20190819" install but not with any other
installs I've tried (neurodebian 7.20190819, 7.20190819 built from
source).

To add even more confusion: With this commit (at least at the run
above [3]), test_AnnexRepo_get_remote_na and test_is_available, which
don't involved SSH and started failing for unknown reasons with this
series and the "conda-forge 7.20190819" install [4], no longer fail.

[4] https://travis-ci.org/github/datalad/datalad/builds/710864799#L3171