BF: status: Pass repo.pathobj-based paths to get_content_annexinfo()
With a repository structured as

    <repo>
    |-- bar -> foo
    `-- foo
        `-- f

some git commands like ls-files return empty when passed the linked
bar/f, while other commands like git-add complain that the path "is
beyond a symbolic link".  In this scenario, git-annex commands take
the no-op/silent route prior to 2a8fdfc7d (Display a warning message
when asked to operate on a file inside a directory that's a symbolic
link to elsewhere, 2020-05-11), but following that commit git-annex
behaves like git-add and errors.

However, unlike git, git-annex also complains when the symlinked
directory is upstream of the top-level directory of the repository and
an absolute path is specified.  This change in behavior results in two
test failures in our test suite for the _DL_TMPDIR="/var/tmp/sym link"
job [0].  These failures have the same underlying cause: status()
calls get_content_annexinfo() with absolute paths that start with
"/var/tmp/sym link", leading to 'git annex find' emitting an error.

Avoid these failures by making status() call get_content_annexinfo()
in a way that is consistent with how it calls diffstatus().  In its
call to diffstatus(), status() converts paths to use
<dataset>.repo.pathobj (in which symlinks have been resolved) instead
of <dataset>.pathobj (in which they have not).

Also, add a new test case for the "symlinked directory within repo"
scenario depicted above, showing how git-annex-find behaves before
8.20200522 (empty result) and after (CommandError).

The change in behavior for linked directories outside of the
repository is probably worth discussing on git-annex's side.  Even if
addressed there, though, the change here is still good to make for the
following reasons.

  * It makes `status` work for the "linked upstream directory" case
    for git-annex >= 8.20200522, with no downside that I'm aware of.

  * `status` already passes these GitRepo.path-recoded paths to
    `diffstatus`.  It then takes the result from that call, which is
    keyed by the recoded paths, and passes it as `init` to
    `get_content_annexinfo`, while passing the original
    Dataset.repo-based paths as the `paths` argument.  When those two
    sets don't match, this happens to not cause any issues (because
    paths come out of 'annex find' as relative and are joined to
    GitRepo.path), but it's still inconsistent and unnecessarily
    confusing.

  * The commands are run with GitRepo.path as the current directory.
    In the upstream symlink case, it's confusing to pass in absolute
    paths that are instead based off of Dataset.path, especially when
    (1) we already have the GitRepo.path-adjusted paths at hand
    because we pass them to `diffstatus` and (2) the keys of the
    get_content_annexinfo() result will be GitRepo.path-based.

[0] https://travis-ci.org/github/datalad/datalad/jobs/710485450#L7806