TST: Add upper git-annex version bound for two resolved SSH hangs
The hang of test_annex_ssh with older OpenSSH versions is addressed by
git-annex's cb74cefde (Fix a hang when using git-annex with an old
openssh 7.2p2, 2020-07-21).

The SSH-related hang of the rsync remote that led to
test_sibling_enable_sameas stalling is, for mysterious reasons,
resolved by 4c9ad1de4, which was a part of the 8.20200720 release.
And, if it resurfaces, it should be taken care of by the changes in
aa492bc65 (Fix a hang when using git-annex with an old openssh 7.2p2,
2020-07-22).