TST: sshconnector: Don't assume SSH target shares home directory
test_ssh_open_close() checks that an 'ls <local home>' call on the
local machine matches an 'ls <local home>' call on the SSH target.
That works with the current Travis setup, but it won't once we switch
to a Docker SSH target.

Instead rely on the resolved temporary test directories being exposed
on the target (something nearly all the SSH-based tests depend on
already).