TST: aggregation: Use resolved path for SSH target
For the current Travis SSH setup, the tests use the localhost as the
SSH target, so the tests can rely on the linked $TMPDIR existing
(e.g., as set up by Travis's _DL_TMPDIR="/var/tmp/sym link" scenario).
The Docker SSH target, however, resolves $DATALAD_TESTS_TEMP_DIR
(falling back to $TMPDIR), ensures that directory exists in the image,
and mounts the resolved local directory to the same directory on the
remote.