Use a fully random UUID as dataset ID
Until now the dataset ID encoded the time of dataset creation, and
a machine-identifying MAC-address (which may or may not be a
reliable/persistent identifier).

DataLad neither used, nor advertised this encoded information in the
ID. Moreover, having such information unconditionally encoded can
counteract other features (e.g. "fake dates").

I'd argue that if identifying information of this kind is desired,
it should be placed explicitly into a dataset. If UUIDs with encoded
content are desired, the UUID of a dataset can be altered (right after
creation) to match such needs.

Fixes gh-4785