TST: clone: Fix test_clone_unborn_head_sub() for managed branch
test_clone_unborn_head_sub() sets up a scenario in which the
to-be-cloned dataset has a submodule where HEAD points to a branch
with no commits and another branch has commits.  This is an uncommon
state to get into with a local repository, but it's not too tricky to
do when setting up remotes, particularly for datasets created with a
create-sibling before 845242677f (ENH: create-sibling: Set HEAD of
sibling to local repo's branch, 2020-04-01).

When on a managed branch, the scenario is botched.  The corresponding
branch is renamed while the adjusted branches are left as is, leading
to a mismatched set like refs/heads/other,
refs/heads/adjusted/master(unlocked), and
refs/basis/adjusted/master(unlocked).  Fix the setup so that the
adjusted branches are renamed as well, following the approach used in
another test, test_clone_unborn_head().

Note that, without the changes here, this test doesn't start to fail
until Git's 7c5045fc18 (ref-filter: apply fallback refname sort only
after all user sorts, 2020-05-03), which was released with v2.27.0.
With 7c5045fc18, the git-annex-init call in the cloned submodule fails
trying to update a non-existent origin/master.  But 7c5045fc18 is just
about refname ordering; you can trigger the failure with Git versions
before v2.27.0 by replacing the refname "other" with "abc"
(i.e. changing the position of the non-master refname relative to
master).

Even though the scenario that triggers this is broken and unlikely to
happen in the wild, this should probably be reported to git-annex in
case it indicates a wider issue.

Fixes #4727.