ENH: Look for dataset procedure in refds on `create`
Supports use cases where a particular configuration needs to be applied
to all subdatasets of a superdataset. It makes sense to put such a
procedure in the root dataset.

This change temporarily expands any existing search location
configuration to include any locations configured in a superdataset
(including the default location).

Fixes gh-4802