NF: Inter-process locking for configuration modifications
As described here:
https://github.com/datalad/datalad/issues/450#issuecomment-670365732
the manipulation of configuration is a collision point for multiple
processes interacting with a single dataset.

This change introduces an inter-process locking mechanism using the
`fasteners` package. Relevant bits from the docs:

- does not require any cleanup
- lock gets dropped automatically if the process crashes, even if
  `__exit__` is not executed

Locking behavior is logged at level 5 and looks like this:

```
[Level 5] Acquired file lock `b'/tmp/datalad_temp_test_obtaink1uf9sf9/.git/config_dataladlock'` after waiting 0.000s [1 attempts were required]
[Level 5] Unlocked and closed file lock open on `b'/tmp/datalad_temp_test_obtaink1uf9sf9/.git/config_dataladlock'`
```

The `fasteners` package is already a direct dependency. It seems to
provide the standard solution for this problem.

Doing locking only at config-write, instead of more global "dataset is
in use"-locks has the advantage of minimizing bottlenecks with external
parallelization approaches (i.e. via `xargs -P`). However, it is
obviously not sufficient in general.

One could debate whether the lock should also cover read operations to
avoid missing partial update done by process A when process B is
reloading a dataset configuration.