TST: test_script_shims: Update for use of console_scripts
As of d40822340 (Use console_scripts entry points, 2020-07-12), the
scripts may contain "EASY-INSTALL-ENTRY-SCRIPT" and an import from
pkg_resources when installed with `pip install -e`.

Guard the two assertions that are no longer valid for editable
installs.  As noted in the comment, the "is editable?" detection is
susceptible to false positives when running the tests from the top of
the source tree, but that's irrelevant for our CI setups, so this
should good enough.

Suggested-by: @yarikoptic