BF+RF: Consolidate ConfigManager on pathlib; handle .git moves
All path operations now use pathlib. The internal variables of
ConfigManager have been reduced in favor of keep a repository
instance.

A local git config file location is now updated on `reload()` to
be able to handle moves of the git-dir, e.g. when fixing up
submodules.

A fix tests were adjusted to actually run on an existing repository
instead of a directory with just files. These tests specifically
verify behavior in 'dataset'-mode, so this seems appropriate.