NF: downloaders: Add support for shub:// URLs
datalad-container resolves shub:// URLs to HTTP URLs by querying
<https://singularity-hub.org/api/container/>.  It then feeds the
result to `git annex addurl`.  That works to download the file, but
the registered URL is short-lived, so it's not useful for retrieving
the image later with `{datalad,git annex} get`.

Add a downloader that resolves shub:// URLs.  When the datalad special
remote is enabled, this allows the shub:// URL, rather than the
resolved one, to be registered with git-annex.

In order for -container to make use of this, containers-add will need
to drop its custom handling when a recent enough datalad is detected.
It should also probably ensure that the datalad special remote is
enabled.

Re: https://github.com/datalad/datalad-container/issues/96