ENH: When closing batched processes avoid waiting forever
This can happen (even though in some cases it is unclear why a
(git-annex) process is not exiting). This patch changes the
unconditional wait to three attempts of 3 seconds with 1 second wait
in between (i.e. 12s). I have no idea whether this is sensible.

Noise log warning announce this condition. If the process hasn't exited
at the end of this period, it is kept running, but is abandoned.

This addresses a stall in crawler tests that was observed in gh-4829,
but did not affect test performance, as it happened during the take-down
phase of the test.

This is also related to the general need for a take-down helper gh-4827