RF: WitlessProtocol: Don't store result in Future object
WitlessProtocol.process_exited() stores the result in the Future
object that run_async_cmd() waits for.  That usually works, but if the
command completes quickly, process_exited() may be called before
pipe_data_received(), in which case the result will not have the
output.

Instead make run_async_cmd() retrieve the result directly from the
protocol.  This makes it possible to add another await that is not
tied to process_exited() call, which will be done in the next commit.
And it probably isn't too strange of an approach because it follows
what's done in the DateProtocol example of the Python documentation.

https://docs.python.org/3/library/asyncio-protocol.html#loop-subprocess-exec-and-subprocessprotocol