BF: run_async_cmd: Work around dropped output of speedy commands
run_async_cmd() waits for a Future that is marked as done in
WitlessProtocol.process_exited() and then collects the output---which
was received via WitlessProtocol.pipe_data_received()---and returns
it.  However, if the command exits quickly, process_exited() may be
called before pipe_data_received(), in which case run_async_cmd()
returns empty output.  Some of our calls to git seem fast enough to
trigger this issue (gh-4773).

This appears to be a bug in asyncio [1] and can be triggered using a
Protocol class that is identical to one in the Python documentation
[2] aside from using a faster command.  Even if this is resolved
upstream, though, we need a workaround for the Python versions that we
support.

Attempt to reliably capture the output by waiting for
`transport._wait()`.  While there may be more proper ways to solve
this, it's the only workaround I've been able to come up.  The
following suggests that it might do the trick:

  * create_subprocess_{exec,shell} don't seem to suffer from this
    problem.  The Process object they return has a .wait method that
    waits for `transport._wait`.  However, those functions use a very
    different protocol than WitlessProtocol, so it's possible that
    that wait isn't the key difference.  Still, it suggests it's a
    safe thing to do.

  * An open PR, gh-4699, switches the ConfigManager over to using the
    WitlessRunner and is triggering a spread of failures that is
    plausibly due to losing output of the quick 'git config' calls.
    Merging the change here into that PR resulted in a passing Travis
    build [3].

  * With the script posted at the Python bug report, I've been unable
    to trigger any dropped output on two runs with 10000 iterations if
    I add `await asyncio.ensure_future(transport._wait())`.  Without
    that line, I've consistently been able to trigger it with many
    fewer iterations.

  * In a reproducer script (included in this PR) that involves Datalad
    and `git commit -- non-existing-file` (inspired by gh-4773), I've
    been unable to trigger the error in hundreds of iterations with
    the changes in this series.  Before these changes, I've been able
    to consistently trigger it using fewer than 100 iterations.

This reverts 2955ba151 (TST: Skip known failure in
test_AnnexRepo_commit, 2020-08-04), as it should resolve the source of
that flaky test failure.

[1] Filed at https://bugs.python.org/issue41594
[2] https://docs.python.org/3/library/asyncio-protocol.html#loop-subprocess-exec-and-subprocessprotocol
[3] https://travis-ci.org/github/datalad/datalad/builds/721449200

Closes #4773