RF: replace wrapped "new_?func" with meaningful "_wrap_<wrappedfunction>"
All the generators and "newfunc"s in the traceback complicate
troubleshooting etc.  I think we got used to this bad practice and
should abandon it and give wrapped functions names reflecting of the
decorator name.  I chose `_wrap_`, since first I chose just `_` and
ran into a conflict interfering with outside `_normalize_path`.
So decided to be more explicit