RF: weakref.finalize instead of __del__
Use a registered callback to clean up for *Repo instances instead of
relying on __del__ being called, which it isn't always. Should be
particularly important for closing BatchedAnnex, which didn't always
happen.

Flyweight tests for *Repo are somewhat flaky on my box ATM. They appear
to not gc correctly when the last strong reference is killed and as a
consequence deliver the same object again when *Repo(somepath) is called
again, which is where it fails (comparing object ids, that should be
different). However, this happens for me only on the first trial to run
those tests. Subsequent calls pass for some weird reason. For now I just
give it a bit time (sleep - yeah), but would be good to figure out what's
actually going on.