BF/TST: Don't use id()
Tests for proving flyweight instance destruction used to be very
unreliable. There are several details leading to confusion in analysis,
but ultimately there was the assumption that a new object would get a
new id. This isn't necessarily true. As the id in CPython is the memory
address it's unique only for a particular point in runtime. Memory can
and will be reused, so later on a new object can turn out to have the
same id a now dead object used to have.
So, proving destruction of an object needs to be a bit more complicated,
but this commit should fix it.