RF: don't create cyclic reference in ConfigManager
ConfigManager only needs to know the relevant paths of a repo it is
bounded to. The *Repo object itself is not needed. Avoid the reference,
since the repo is referencing its ConfigManager already. Without the
circle we should be able to rely solely on refcount based garbage
collection, which absolutely is preferrable.