NF: Minimalist snakemake wrapper
It injects a `datalad-get` call in front of every file property
inspection by snakemake, and thereby should ensure that the files in a
dataset that are relevant for a workflow are actually around for a valid
inspection.

This approach does not deal with the use case where a file can be
downloaded, but is actually cheaper to recompute.