MNT: gitrepo: Warn if problematic Git version is detected
The master branch (what will be 0.14) already contains the previous
commits in this series, declaring a minimum git version of 2.19.1 and
aborting if we detect something below it.  As described in ff042b88a
(ENH: gitrepo: Declare and check minimum Git version, 2020-06-22),
this minimum version is forward-looking and based off what's available
from particular distributions.  The first Git release with a known
issue is actually much earlier, v2.13.

Merging this series to the maint branch as is then would be too
aggressive for a maintenance release because no prior release has
declared a minimum version, yet we'd be making DataLad non-functional
with a range of Git versions that don't have known issues.  And even
raising an error if Git is below v2.13 could be problematic because
the subset of functionality that a user relies on may not be affected.

Instead issue a warning if the detected Git is below v2.13.0 to let
users that run into problems know that their Git version is likely the
culprit.

Re: #4650