TST: get: Add test for direct fetch of subdataset commit
This series drops get's use of 'git submodule update --init'.  Like
'git submodule update --init', our custom logic tries to fetch a
subdataset commit directly if it's not present.  Add a test for that.

Note that SSH is needed because, for a local path, git-clone's --local
behavior would kick in and the commit would be available.