TST: Refcount is unreliable ATM
Due to issue #4862, we currently still require gc.collect() under unclear
circumstances to get rid of an exception traceback when creating in an
existing directory. That traceback references the respective function
frames which in turn reference the repo instance (they are methods).
Doesn't happen on all systems, though. Eventually we need to figure that
out.
However, still test for the refcount after gc.collect() to ensure we
don't introduce new circular references and make the issue worse!