TST: repodates: Adjust logic for synced adjusted branches
test_check_dates() makes two sets of commits, one before the reference
date and one after, and then checks that the commit is reported in the
older and newer output, respectively, of repodates.check_dates().

With 38b79ee61 (RF: Adjust tests for Repo.add() removal), the check
for the older commit started to fail.  This is because calling
repo.save(), unlike calling repo.add(), leads to a localsync() call
that effectively rebases the adjusted branch.  As a result, the
original commit is no longer reachable from a local branch, and
repodates.check_dates() does not pick it up when called with revs=None
(which says to check commits in all local branches).

Instruct repodates.check_dates() to check all refs so that we pick up
the older commit, which is still reachable from an annotated tag and
lightweight tag.