BF+TST: ephemeral clone to account for bare repo
In context of RIA stores it's important to be able to have an ephemeral
clone of a bare repository. clone --reckless=ephemeral didn't account
for that and symlinked into '.git/annex' of a bare repo, which is
invalid, of course.

Added test for correct symlinking of that clone option and another one
for an actual workflow involving a RIA store.

(Closes #4893)