MNT: treewide: Consistently use ensure_ variants from utils
gh-3909 renamed each assure_* function in datalad.utils to use
ensure_, assigning the new names to the old ones for compatibility.
We use the assure_ variants widely throughout our code base.  At that
time, we decided to temporarily let those occurrences be to avoid
merge conflicts.

Do that bulk replacement now.

Note that there are still a couple of assure_ methods in the code
base.  Those are left to another series, if someone thinks those are
worth cleaning up.