RF: Change ConfigManager.get() behavior for multiple identical key specs
Before this change, the values of all specifications of an identical key
were returned as a tuple. Now only the last value is returned, and
`get_all=True` has to be set to get the previous behavior.

Rationale: That a single value is returned is a sensible default
expectation. Valid use cases for multiple values are few, and require
custom handling in consuming code. The previous behavior would require
protection against the multi-value-return possibility all throughout
the code base. This switch is disruptive, but makes overall operation
more robust.

Fixes gh-2628
Fixes gh-4919