BF: use time.time (no timezone offset) while checking expiration for AWS S3 credential
calendar.timegm expects the time tuple to be in GMT time tuple.
time.localtime() has time zone information but it is not part of the time tuple.

calendar.timegm then produced seconds since epoch for that in GMT thus
zone information was not used.

With time.time() we seems to be getting seconds since epoch properly.

I kept running into S3 downloads failing while credential reported to
be expiring in some hours to come