MNT: Drop now unneeded handling of config.get() return value
With the last commit, config.get() now returns a single value by
default instead of a tuple in the case of multiple values for the same
key.  Adjust the config.get() callers that detect a tuple and pull out
the last value, as this is now what config.get() does by default.