BF: config: Rewrite get<type>() methods to use get()
The config get*() methods return a tuple if there are multiple values
for the same key (in the non-committed git configuration, if the key
is present there, or in the dataset configuration).  Very few get*()
callers in the code base account for the possibility of a tuple return
value.

With the first commit of this series, get() now returns the last value
of the tuple by default, following the behavior of `git config <key>`.
This eliminates the above problem for get() callers, but it doesn't
cover get_value() or the convenience methods that go through
get_value(): getint(), getfloat(), and getbool().

get_value() probably isn't worth adjusting; it isn't widely used and
it exists for historical reasons (compatibility with the now
irrelevant GitPython parser).  Aside from splitting the `key`
parameter into `section` and `option`, get_value() differs from get()
in one way: if `default` is None, a KeyError is raised rather than
returning None.

getbool(), on the other hand, is used a good amount in non-test code,
and it seems that all the callers expect the single value behavior of
`git config <key>`, as none account for the exception that would
happen in the tuple case (see gh-2628).  Adjust getbool() to go
through get() for the tuple-to-single-value handling, while keeping
the 'default=None -> KeyError' handling for backward compatibility.
For completeness, do similar for getfloat() and getint(), though
combined these are only used in one non-test spot.

Re: #4923
Fixes #2628