BF: Expect multi-value config keys in cfg_metadatatypes procedure
Previous implementation was suboptimal/wrong as it did not ensure
the "X in Y" test is happening with Y always being a tuple/list.

Now also uses the new ConfigManager.get(get_all=True) option to
explicitly request all values to be reported.

Adjusted a related test to trigger multi-value issues in the -core
tests, instead of solely relying on an associated -neuroimaging test
as an indicator.