ENH(+RF): @try_multiple_dec - add exceptions_filter + doc, mv msg where it is correct
Main motivation -- I keep breeding custom loops (e.g.
https://github.com/datalad/datalad/pull/4928 ) or similar situations.

Adding an "exceptions_filter" argument would allow to make it flexible
to fit more use cases, where decision making should be made based on
details of the exception (e.g. .status for S3ResponseError).

Also I moved log message which promises that we are about to sleep
right before the sleep, so we do not log that if we are about to
re-raise it instead.