RF: remove disable_logger in customremotes/datalad.py
It was introduced in 15ecbb78a86f36b353e763273577dd369671e61b (0.9.2~76^2~6)
as a part of https://github.com/datalad/datalad/pull/1870
with a purpose according to description of not pollutting stdout used by
special remote for communication.  BUT logs go to stderr and do not interfer
anyhow with the communication, so the reason for the change is not clear to me.

I have been trying to figure out why the heck I am still seeing failures after
https://github.com/datalad/datalad/pull/4931 .  I kept adding more logging but
nothing appeared in the logs!  I finally was brought to this piece of code,
thus the motivation and argumentation for the change.