Revert "RF: make needs_authentication into a bool"
This reverts commit c936ed37d29b5175d49f27e020cae3498a9c5c92.

I think there was some thought behind initial code since "else" clause does
cause entry of credential if there is a no known credential and we do not
know by then (thus None) if needs_authentication (would be bool) if there is no
authenticator assigned.

So to not disturb that portion of the code, I am reverting the change
which was incomplete anyways since logic in "else:" would no longer trigger
nested else: to entry of credential.