ENH+RF: CompositeCredential - add .refresh + trigger regeneration of full chain upon enter_new
We need interface to "force-trigger" regeneration of apparently (but
not due to expiration datetime as we know) credential.  .refresh()
does that and is now used by enter_new().

Change in behavior is that we will now trigger full chain to get all
"tail" credential, but I take it as a feature -- so we do not delay
with generation and thus possibly surface code bugs etc only until
used.  Also, it would make `is_expired` assessment valid - since if we
do not regenerate, is_expired ATM would not be adequately report
either some old generated credential is expired or not.

A possible disadvantage -- it typically would require network
connection while simply entering a new credential. But IMHO it is Ok