RF+ENH: downloaders (s3) - raise and handle dedicated AccessPermissionExpiredError
s3 error_codes MIGHT (see comment -- we do not always get them) hint
us on the nature of 400 code.  If it is ExpiredToken, we need to
handle it more specifically than just "keep trying" and then rely on
our knowledge of when it is to expire.

To say the truth -- I have triggered this use specific use case
manually: I have replaced proper token with an expired one, which in
theory should not happen.  But who knows -- may be admins would
explicitly expire some tokens which theoretically should still be
valid (according to initial expiration date), so we better be able to
handle such scenarios as well instead of just demanding a new full
credential to be entered.