RF+ENH: centralize check of credential not expired and do before .get_key()
This way we would not even attempt (immediately after) getting a key if
we know that credential is expired already.   There can still be a race condition and proper reaction to the expiration of the token is yet to be figured out
how to handle properly while handling 400

I am also thinking may be about "artificially" bringing (stored) expiration
slightly closer in time, e.g. making it 1% closer than the whoe allotment.
That would help to avoid any possible race as long as it is to expire just in a
few seconds since allotment... or may be change it to be 2 seconds closer
altogether -- unlikely expiration would ever be that close in time!