ENH: a non-excercised integration test to verify correct timing out of tokens for s3
good news: it passes if ran locally.  Bad news: abcd addurls failed again with 400
and from all the logs it looks that it might no longer be timed out token
(it connects find for about 2 minutes before it kabooms with 400).  I was hoping
that it is still our code somehow but test passes so I will assume that it is
something else

I have decided to commit this "dead code" to ease possible troubleshooting later on