BF: do not urlunquote filename of a file from an archive
Wow -- this must be one of the oldest buggy lines of code in the datalad from the
times when both earth and moon were flat: initial commit which introduced this unquote
is 0.0.1-48-g2170d1477 (2170d14778e9ca1bf7df17a24ee0ca9ce3326e79).

Commit message there
 ENH/RF: Special remote provided optional path

though hints on possibility of handling unquoting somewhere in the special remotes code.
We will see if that is needed

Closes gh-4950