BF: addurls: Fix subpath sorting type error
As noted by @yarikoptic (and also the failing tests), top-level
datasets have a subpath of None, so we can't sort on the subpath key
directly.

Fall back to an empty string when sorting.  Another option would be to
adjust the upstream code, in particular _format_filenames(), to use an
empty string instead of None for the top-level dataset's subpath.  Go
with the more local change for now.

https://github.com/datalad/datalad/pull/4973#discussion_r498894503