Workaround: Add list of common textfiles on Windows in text2git
As Mimeencoding is not yet functional on Windows (see #3360), the text2git configuration is broken,
and leaves the dataset in an unusable state once applied (until the configuration is removed by hand).
This change sets largefile = nothing rules for a list of common text file extensions when on Windows
instead of the standard text2git configuration.

Note that this means that datasets created with text2git on non-Windows machines aren't portable to
Windows, but vice-versa.