RF/BF: report "impossible" upon get of a annexed file without git-annex branch
Solution here is incomplete (comment added) since it should consider git
link files in cases of unlocked annexed files.  It would be more expensive AFAIK,
thus before even attempting to implement that, I have decided to provide the
fix which would work in 90% of the time first.

Specifically it would Closes #4510

Without the fix, asking get to get content for a broken symlink results in "ah
-- all is good -- you do not need it" behavior (similar to the slogan of the
local little town store "We do not have it - you do not need it!"), which is
IMHO plain incorrect:

	$> datalad -f json_pp get sub-01/anat/sub-01_inplaneT2.nii.gz
	{
	  "action": "get",
	  "path": "/tmp/ds000001/sub-01/anat/sub-01_inplaneT2.nii.gz",
	  "refds": "/tmp/ds000001",
	  "status": "notneeded",
	  "type": "file"
	}

with exit 0.  With the fix:

	$> datalad -f json_pp get sub-01/anat/sub-01_inplaneT2.nii.gz
	[WARNING] no dataset annex, content already present [get(/tmp/ds000001/sub-01/anat/sub-01_inplaneT2.nii.gz)]
	{
	  "action": "get",
	  "path": "/tmp/ds000001/sub-01/anat/sub-01_inplaneT2.nii.gz",
	  "refds": "/tmp/ds000001",
	  "status": "impossible",
	  "type": "file"
	}
	(dev3) 1 29415 ->1

exits with 1, and status "impossible" which is IMHO factually correct, and
WARNING message is actually informative.