BF: Don't modify PATH if not necessary
PATH modification in Runner classes is meant to make sure we use
bundled w/ annex git executable (except when instructed otherwise
by DATALAD_USE_DEFAULT_GIT). However, don't mess with PATH when
what we find to be bundled is the same thing that is found with
unmodified PATH anyway. This is particularly important when the
'bundeled' location is /usr/bin and we are running in a virtual
environment, leading to that PATH overruling the venv.

(Closes #5030)