RF: MultiplexSSHConnection(BaseSSHConnection)
Make room for a representation of a non-multiplexed SSH connection.
Anticipates a leaner SSHManager too, which does not need to keep track
of single-process SSH connections.

Also streamline the handling of SSH args. Instead of keeping individual
ones, let the constructor(s) lump them into categories "for opening" and
"for any call". This seems to be sufficient.

Ping #2575